requirements.txt
alembic==1.13.3
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.6.2
asttokens==3.0.0
attrs==25.4.0
bcrypt==4.0.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
decorator==5.2.1
Deprecated==1.2.18
distlib==0.4.0
ecdsa==0.19.1
executing==2.2.0
Faker==37.8.0
fastapi==0.114.0
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==4.0.0
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
flask-restx==1.3.2
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httptools==0.7.1
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
ipdb==0.13.13
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
limits==5.6.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pipenv==2025.0.4
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlalchemy-serializer==1.4.22
stack-data==0.6.3
starlette==0.38.6
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.23.1
uvloop==0.21.0
virtualenv==20.34.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.2
wrapt==1.17.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
sendgrid==6.11.0
gunicorn==23.0.0
flask
flask-restx
flask-jwt-extended
flask-sqlalchemy
flask-migrate
python-dotenv
flask-cors
psycopg2-binary
requests
pytest
pytest-flask
pytest-mock
flask-marshmallow

flask-socketio
